pyproject.toml
[project]
name = "vireka-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "catboost>=1.2.8",
    "certifi>=2025.4.26",
    "earthengine-api>=1.5.17",
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "numpy>=2.2.6",
    "openai>=1.82.1",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "supabase>=2.15.2",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
